Initial commit. Modified some dispatches cause of a type alias.

+++ src/nearneighbors.jl
@@ -15,13 +15,13 @@ const BoolVal = Union{Val{true},Val{false}}
 @inline neighbor_info_type(x) = neighbor_info_type(typeof(x))
 
 # Neighborhoods
-const NeighborInfo{X<:SampleIndex,D,U} = NamedTuple{(:index,:cost,:controls),Tuple{X,D,U}}
+const NeighborInfo{X<:SampleIndex,D,U} = NamedTuple{(:index, :cost, :controls),Tuple{X,D,U}}
 const Neighborhood{X<:SampleIndex,D,U} = AbstractVector{NeighborInfo{X,D,U}}
-includes_controls( ::Type{  NeighborInfo{X,D,U}}) where {X,D,U} = U !== Missing
-includes_controls( ::Type{<:Neighborhood{X,D,U}}) where {X,D,U} = U !== Missing
+includes_controls(::Type{NeighborInfo{X,D,U}}) where {X,D,U} = U !== Missing
+includes_controls(::Type{<:Neighborhood{X,D,U}}) where {X,D,U} = U !== Missing
 neighbor_info_type(::Type{<:Neighborhood{X,D,U}}) where {X,D,U} = NeighborInfo{X,D,U}
-indextype(   ::Type{NeighborInfo{X,D,U}}) where {X,D,U} = X
-costtype(    ::Type{NeighborInfo{X,D,U}}) where {X,D,U} = D
+indextype(::Type{NeighborInfo{X,D,U}}) where {X,D,U} = X
+costtype(::Type{NeighborInfo{X,D,U}}) where {X,D,U} = D
 controlstype(::Type{NeighborInfo{X,D,U}}) where {X,D,U} = U
 function _neighbor_info_type(::Type{X}, ::Type{NamedTuple{M,Tuple{D,U}}}, include_controls::BoolVal) where {X,M,D,U}
     include_controls === Val(true) ? NeighborInfo{X,D,U} : NeighborInfo{X,D,Missing}
@@ -35,10 +35,10 @@ function shift_lattice_component(n::NeighborInfo{<:TiledIndex}, shift)
     (index=shift_lattice_component(n.index, shift), cost=n.cost, controls=n.controls)
 end
 function neighbor_info_iterator(nodes, bvp, x, inds, r=nothing; dir::F_or_B=Val(:F),
-                                include_controls::BoolVal=Val(false), omit::F=always_false) where {F}
+    include_controls::BoolVal=Val(false), omit::F=always_false) where {F}
     ((index=i, keep_controls_if(dir === Val(:F) ? (r === nothing ? bvp(x, nodes[i]) : bvp(x, nodes[i], r)) :
-                                                  (r === nothing ? bvp(nodes[i], x) : bvp(nodes[i], x, r)),
-                                include_controls)...) for i in inds if !omit(i))
+                                (r === nothing ? bvp(nodes[i], x) : bvp(nodes[i], x, r)),
+        include_controls)...) for i in inds if !omit(i))
 end
 function reverse_neighbor_dict(i, neighborhood, ::Type{NeighborInfo{X,D,U}}) where {X,D,U}
     Dict{X,NeighborInfo{X,D,U}}(n.index => (index=i, cost=n.cost, controls=n.controls) for n in neighborhood)
@@ -57,14 +57,14 @@ function PackedNeighborhoods(neighborhoods::AbstractVector{<:Neighborhood})
 end
 Base.length(PN::PackedNeighborhoods) = length(PN.neighborhoods)
 Base.getindex(PN::PackedNeighborhoods, i::Int) = PN.neighborhoods[i]
-includes_controls( ::Type{<:PackedNeighborhoods{X,D,U}}) where {X,D,U} = U !== Missing
+includes_controls(::Type{<:PackedNeighborhoods{X,D,U}}) where {X,D,U} = U !== Missing
 neighbor_info_type(::Type{<:PackedNeighborhoods{X,D,U}}) where {X,D,U} = NeighborInfo{X,D,U}
 
 # k-Nearest Neighbors
 struct Nearest
     k::Int
 end
-placeholder_neighborhood(::Type{T}, k) where {T<:NeighborInfo} = Vector{Union{Missing,T}}(missing, k)
+placeholder_neighborhood(::Type{T}, k::Int) where {T<:NeighborInfo} = Vector{Union{Missing,T}}(missing, k)
 function update_knn!(neighborhood, n::NeighborInfo{X,D,U}) where {X,D,U}
     if neighborhood[1] === missing || n.cost < neighborhood[1].cost
         neighborhood[1] = n
@@ -77,7 +77,7 @@ end
 abstract type NeighborhoodCache end
 includes_controls(::Type{<:NeighborhoodCache}) = false
 function neighbors(cache::NeighborhoodCache, nnds, nodes, bvp, v_or_x, r;
-                   dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false))
+    dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false))
     neighbors(nnds, nodes, bvp, v_or_x, r, dir=dir, include_controls=include_controls)
 end
 
@@ -101,19 +101,19 @@ function MemoizedNC(::Type{T}, n::Int=0, issymmetric::Bool=false) where {T}
 end
 function MemoizedNC(nodes::SampleSet, bvp::SteeringBVP; include_controls::BoolVal=Val(false))
     MemoizedNC(neighbor_info_type(nodes, bvp, include_controls=include_controls),
-               length(nodes), issymmetric(bvp) && include_controls === Val(false))
+        length(nodes), issymmetric(bvp) && include_controls === Val(false))
 end
-includes_controls( ::Type{MemoizedNC{N}}) where {N} = includes_controls(N)
+includes_controls(::Type{MemoizedNC{N}}) where {N} = includes_controls(N)
 neighbor_info_type(::Type{MemoizedNC{N}}) where {N} = neighbor_info_type(N)
 reset!(cache::MemoizedNC) = (fill!(cache.neighborhoods_F, missing); fill!(cache.neighborhoods_B, missing))
 function neighbors(cache::MemoizedNC{N}, nnds, nodes, bvp, v::SampleIndex, r;
-                   dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false)) where {N}
+    dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false)) where {N}
     neighborhoods = dir === Val(:F) ? cache.neighborhoods_F : cache.neighborhoods_B
     i = linear_index(v)
     neighborhood = i > 0 ? neighborhoods[i] : missing
     if neighborhood === missing
         neighborhood = collect_if_not_Array(neighbors(nnds, nodes, bvp, v, r,
-                                                      dir=dir, include_controls=include_controls))
+            dir=dir, include_controls=include_controls))
         i > 0 && (neighborhoods[i] = neighborhood)
     end
     neighborhood
@@ -132,10 +132,10 @@ function OfflineNC(cache::MemoizedNC)
         OfflineNC(PackedNeighborhoods(cache.neighborhoods_F), PackedNeighborhoods(cache.neighborhoods_B))
     end
 end
-includes_controls( ::Type{OfflineNC{PN}}) where {PN} = includes_controls(PN)
+includes_controls(::Type{OfflineNC{PN}}) where {PN} = includes_controls(PN)
 neighbor_info_type(::Type{OfflineNC{PN}}) where {PN} = neighbor_info_type(PN)
 function neighbors(cache::OfflineNC{PN}, nnds, nodes, bvp, v::SampleIndex, r;
-                   dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false)) where {PN}
+    dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false)) where {PN}
     dir === Val(:F) ? cache.neighborhoods_F[linear_index(v)] : cache.neighborhoods_B[linear_index(v)]
 end
 
@@ -150,20 +150,20 @@ use_NullNNDS() = (default_NN_data_structure_type[] = :NullNNDS)
 reset!(::NullNNDS) = nothing
 addstates!(::NullNNDS, X) = nothing
 function neighbors(nnds::NullNNDS, nodes, bvp, v_or_x, r;
-                   dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false))
+    dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false))
     neighbors(nodes, bvp, v_or_x, r, dir=dir, include_controls=include_controls)
 end
 
 ## MemoizedNNDS
 struct MemoizedNNDS{S,D,U,F} <: NearNeighborDataStructure
-    bvp_results::Dict{Tuple{S,S},NamedTuple{(:cost,:controls),Tuple{D,U}}}
+    bvp_results::Dict{Tuple{S,S},NamedTuple{(:cost, :controls),Tuple{D,U}}}
     memoized_bvp::F
 end
 function MemoizedNNDS(nodes::ExplicitSampleSet, bvp::SteeringBVP; include_controls::BoolVal=Val(false))
     S = typeof(nodes.init)
     D = typeof(bvp(nodes.init, nodes.init).cost)
     U = include_controls === Val(true) ? typeof(bvp(nodes.init, nodes.init).controls) : Missing
-    bvp_results = Dict{Tuple{S,S},NamedTuple{(:cost,:controls),Tuple{D,U}}}()
+    bvp_results = Dict{Tuple{S,S},NamedTuple{(:cost, :controls),Tuple{D,U}}}()
     memoized_bvp = (x0, xf, cost_bound=D(Inf)) -> begin    # TODO: check if closure performance is disastrous
         get!(bvp_results, (x0, xf)) do                     #       also, make subtype of SteeringBVP?
             keep_controls_if(bvp(x0, xf, cost_bound), include_controls)
@@ -176,23 +176,24 @@ includes_controls(::MemoizedNNDS{S,D,U}) where {S,D,U} = U !== Missing
 reset!(nnds::MemoizedNNDS) = empty!(nnds.bvp_results)
 addstates!(::MemoizedNNDS, X) = nothing
 function neighbors(nnds::MemoizedNNDS, nodes, bvp, v_or_x, r;
-                   dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false))
+    dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false))
     neighbors(nodes, nnds.memoized_bvp, v_or_x, r, dir=dir, include_controls=include_controls)
 end
 
 # General Methods
 function neighbors(nodes::ExplicitSampleSet, bvp, v::Int, r;
-                   dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false))
+    dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false))
     neighbors(nodes, bvp, nodes[v], r, dir=dir, include_controls=include_controls, omit=isequal(v))
 end
 function neighbors(nodes::ExplicitSampleSet, bvp, x::State, r;
-                   dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false), omit::F=always_false) where {F}
+    dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false), omit::F=always_false) where {F}
     itr = neighbor_info_iterator(nodes, bvp, x, eachindex(nodes), r,
-                                 dir=dir, include_controls=include_controls, omit=omit)
+        dir=dir, include_controls=include_controls, omit=omit)
     Iterators.filter(n -> n.cost <= r, itr)
 end
+# If r = Nearest, then r=missing in neighbor_info_iterator
 function neighbors(nodes::ExplicitSampleSet, bvp, x::State, r::Nearest;
-                   dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false), omit::F=always_false) where {F}
+    dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false), omit::F=always_false) where {F}
     itr = neighbor_info_iterator(nodes, bvp, x, eachindex(nodes), dir=dir, include_controls=include_controls, omit=omit)
     neighborhood = placeholder_neighborhood(neighbor_info_type(nodes, bvp, include_controls=include_controls), r.k)
     foreach(n -> update_knn!(neighborhood, n), itr)
@@ -200,7 +201,7 @@ function neighbors(nodes::ExplicitSampleSet, bvp, x::State, r::Nearest;
 end
 
 function neighbors(nodes::TiledSampleSet, bvp, v::TiledIndex, r;
-                   dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false))
+    dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false))
     if v.off_lattice <= 0
         neighbors(nodes, bvp, nodes[v], r, dir=dir, include_controls=include_controls)
     else
@@ -213,8 +214,8 @@ function neighbors(nodes::TiledSampleSet, bvp, v::TiledIndex, r;
     end
 end
 function neighbors(nodes::TiledSampleSet, bvp, x::State, r;
-                   dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false), omit::F=always_false,
-                   x0_xf=(dir === Val(:F) ? i -> (x, nodes[i]) : i -> (nodes[i], x))) where {F}
+    dir::F_or_B=Val(:F), include_controls::BoolVal=Val(false), omit::F=always_false,
+    x0_xf=(dir === Val(:F) ? i -> (x, nodes[i]) : i -> (nodes[i], x))) where {F}
     neighborhood = neighbor_info_type(nodes, bvp, include_controls=include_controls)[]
     close_idx = findclose(nodes, bvp, x, r, dir=dir, omit=omit, x0_xf=x0_xf)
     queue = [close_idx.on_lattice]
@@ -240,9 +241,9 @@ function neighbors(nodes::TiledSampleSet, bvp, x::State, r;
     neighborhood
 end
 function findclose(nodes::TiledSampleSet, bvp, x::State, r; dir::F_or_B=Val(:F), omit::F=always_false,
-                   x0_xf=(dir === Val(:F) ? i -> (x, nodes[i]) : i -> (nodes[i], x)), cutoff::Int=30) where {F}
+    x0_xf=(dir === Val(:F) ? i -> (x, nodes[i]) : i -> (nodes[i], x)), cutoff::Int=30) where {F}
     Q, R = qr(nodes.A)
-    start_lattice_idx = round.(Int, R\Q'*(x - nodes.b))
+    start_lattice_idx = round.(Int, R \ Q' * (x - nodes.b))
     sentinel = fill(typemin(Int), typeof(start_lattice_idx))
     queue = [start_lattice_idx, sentinel]
     visited = Set(queue)
@@ -252,7 +253,7 @@ function findclose(nodes::TiledSampleSet, bvp, x::State, r; dir::F_or_B=Val(:F),
         if on_lattice_idx == sentinel
             push!(queue, sentinel)
             ct += 1
-            ct > cutoff && error("Either r is too small, or searching near the starting "*
+            ct > cutoff && error("Either r is too small, or searching near the starting " *
                                  "state isn't working to find a close tiled sample!")
             continue
         end
